Change "observation" to "argument"
This change from "observation" to "argument" is justified a few ways.
First, I was unhappy with the word "observation" as being both
too general and too vague, and so sought a new description.
"Argument" more accurately captures what a process is doing, or
what is going on within a process, than does "observation."

Second, if we choose to define a process as a function, then we
must express it as mapping a set of elements to a set of elements;
in this case, the "argument" of a process (or function) must be
well-defined.

Third, as a result of this new description, the entity (process)
becomes better-defined-- and, I feel, still retains the accuracy
of the intended meaning.

Fourth, as a matter of coincidence, this new view facilitates
viewing the process as a whole-- that is, viewing all three
parts of a process simultaneously, rather than viewing them
in succession.

The change from "resolution" to "result" is the compliment of
this change.