Introduced "potential change"
Deleted the sentence "Intuitively, a process makes a change."
from the introduction; replaced with
"All processes make change.  A process is viewed as a potential
change, or, as a change potential.

First, it just sounds right.
Upon closer inspection, we aren't talking about a literal or actual
change, but rather an abstraction of change, or 'the potential for
change'.
From wikipedia, "Potential generally refers to a currently
unrealized ability."

It might be the case that, this development so far considers, strictly,
future processes (and therefore potential/unrealized change);
that is, there might be 'past change.'
Not being well-grounded in physics, I do not have the education to
assert the validity of Wikipedia's definition first-hand.
Anyways, I just like it.