Forgot the period.
First, the beauty of introducing "realized" processes, as the term is
defined, is that it captures (serves as a suitable reference to) all
instances of the intuitively-defined notion of "realized processes."
As defined, it is more general than that, but so what?

Second, this split between "realized" and "potential" change
allows us to describe well two different aspects of the same thing.